pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
profile = "black"

[tool.poetry]
authors = []
description = ""
name = "django-structured"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
click = "^8.1.7"
click-option-group = "^0.5.6"
python = "^3.12"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ipdb = "^0.13.13"
ipython = "^8.24.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[tool.poetry.group.test.dependencies]
coverage = "^7.5.1"
pytest = "^8.2.0"
pytest-clarity = "^1.0.1"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-randomly = "^3.15.0"
pytest-sugar = "^1.0.0"
pytest-xdist = "^3.6.1"
pyyaml = "^6.0.1"
pytest-subtests = "^0.12.1"

[tool.poetry.scripts]
django-structure = "django_structured.entrypoints:structured"
